Fixes a crash in ElementDecoration class, where deleted elements were being rendered after their memory was freed. The original fix was created by Kezeali on May 20, 2012, located here: https://github.com/Kezeali/libRocket/commit/ea8aea9ee0c6dd4b4d9dcbb564a750548f2165f2
Unfortunately, Kezeali was aiming to solve a different problem which I fixed in the previous commit (still waiting for libRocket to commit it and close the bug report), and therefore his commit has far too much unrelated code to evaluate to even be considered as an official commit. This, much smaller commit, borrows on Kezeali's code exclusively (nothing added by me) and effectively resolves the issue.