Remove a performance optimization to improve thread and exception safety.
The format routine was using an internal static buffer presumably in order to reduce dynamic memory allocations and hopefully gain a performance win. Though noble, this optimization comes at the expense of safety.

IMO a user of such a common and seemingly harmless routine as a formatting function would not expect to loose thread safety just to do string formatting. That fails the test of least surprises.

On that basis, I have removed the optimization to prefer thread and exception safety instead.

To compensate a little I have doubled the default buffer size at which dynamic memory would kick in on the assumption that most situations can reasonably expect to take 2k of the stack temporarily as much as 1k without likely hitting a stack overflow. If that assumption is considered unsound for all contexts that need support, the size change can be reverted independently, but it shouldn't affect the rationale to keep the rest of the change.

If the overall change produces an actual performance issue (as in unacceptable not just slower), it might be better to review the issue in the context of that app where it is unacceptable and if any further optimization would consider reverting this change to something that is unsafe again, it might be better to make that change specific to the app that is having the problem rather than doing it in a generally library like here.